pdx86-cherry: Check for previously applied commits
If the for-next or fixes branches have been rebased (which should only
happen if Linus rejects a pull a request or if linux-next detects a bug
we missed), it's possible they contain previously merged commits with
new commit ids. We need to catch these before sending a pull request.

This script will compare master with the passed in commit id and use
git-cherry to find identical patches by diff, returning an error if
detected.

Signed-off-by: Darren Hart <dvhart@infradead.org>