mutex: Create orphan mutex branch and LOCK file
The new orphan mutex branch contains the LOCK file, identifying which of
the maintainers currently has a lock on the repository. The intent is to
make it clear to the other maintainers that patches are being actively
collected and will be pushed soon. When the LOCK is released, the
branch HEADs will be different, so any work another maintainer may be
doing would need to be rebased before being pushed.

LOCK states:

locked: "dvhart", etc.
unlocked: ""

Signed-off-by: Darren Hart <dvhart@infradead.org>