Add README.md
This is an only slightly cleaned up version of the email I sent to Andy
early on as an introduction to the Platform Drivers X86 tree, maintainer
model, tooling, goals, etc.

I'm including it in its entirety for the initial commit, partially
because that's all I have time for right now and I want to get this
going.

I'll follow this up by removing the fluff and obsolete bits and boiling
it down to current process as we understand it.

Signed-off-by: Darren Hart <dvhart@infradead.org>