pdx86-lock: Add a status command
Add a status command, -s, which will report on the state of the lock,
but not attempt to acquire or free it.

Use the owner command to report out status, and simply test $VAL when we
need to know if we are the owner, rather than having the owner function
perform both and deal with a more convoluted return path.

Signed-off-by: Darren Hart (VMware) <dvhart@infradead.org>