pdx86-tag: Prompt to continue with tag including a duplicate
We will occasionally include a cherry-picked patch from fixes in a tag,
such as when a fix goes in to v4.12-2, but is needed for subsequent
development in the for-next branch being prepared for v4.13.

Add a prompt, defaulting to N, permitting continuing to create the tag
after git cherry identified a duplicate.

Linus's preference for this over merging of the fixes branch is
discussed here:

https://www.spinics.net/lists/kernel/msg2492566.html

We should keep an eye on the times we have to do this and check if
merging fixes would result in pulling in more than the cherry-picked
patches and our own patches, if not, merging fixes after they are pulled
may be preferable.

Signed-off-by: Darren Hart (VMware) <dvhart@infradead.org>